requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb-client==0.3.27.dev0
discord.py==2.3.0
frozenlist==1.3.3
idna==3.4
monotonic==1.6
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
overrides==7.3.1
pandas==2.0.2
posthog==3.0.1
pydantic==1.10.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
